requirements.txt
torch~=2.6.0+cu118
torchvision~=0.21.0+cu118
tqdm~=4.67.1
pandas~=2.2.3
numpy~=2.1.2
matplotlib~=3.10.1
seaborn~=0.13.2
scikit-learn~=1.6.1
wandb~=0.19.9
plotly~=6.0.1
ipython~=8.35.0
pillow~=11.0.0
timm~=1.0.15
opencv-python~=4.11.0.86
albumentations~=2.0.5
mmpretrain~=1.2.0
ranger21~=0.1.0